Update the express-validator version to 6.14
