+++ package.json
@@ -23,7 +23,7 @@
     "config": "^3.3.8",
     "express": "^4.18.1",
     "express-async-handler": "^1.2.0",
-    "express-validator": "^5.3.1",
+    "express-validator": "^6.14.2",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.6.4",
     "nodemon": "^2.0.20",
